@Specializes and @Priority on DynamicJacksonJsonProvider.
Use @Specializes and @Priority annotations so that dependency injection system prioritizes the DynamicJacksonJsonProvider over other (default) providers of the same service, which is present in the container environment.

Without this, sometimes the default provider that does not enable dynamic object mapper lookup (and disables caching) would take precedence, and so the context dependent object mapper resolver would not work as intended.