Add TimeTypesModelConverter, to get a correct openapi spec generated for java.time.Duration properties.
Before this, the openapi spec generated for properties of type java.time.Duration was an object with lots of internal properties, which did not match the (de)serialization actually done by the server.

This change fixes that, so that Duration properties become string properties, with format "duration".

The Duration is (de)serialized from/to string with ISO 8601 format, so this will then be a correct openapi spec.