Fix eternal recursion in resolve.
If the DTO type that was analyzed had a Schema annotation already, it caused an eternal recursion until stackoverflow, because the check for processedTypes.contains in ModelCoverterContextImpl never matched because the RegisteredSubtypesModelConverter always created a new unique type with new ctxAnnotations.

Fixed by not creating a new ctxAnnotations array when the incoming ctxAnnotations already has the same content.