Don't fail when include = JsonTypeInfo.As.EXTERNAL_PROPERTY is used.
When EXTERNAL_PROPERTY is used, and `@JsonTypeInfo` annotation is set on the property declaration for a type, we don't need to make a discriminator, the "container" type has the discriminator on a separate property, so it will be included in the generated typescript. But it also means the generated typescript will be a simple union type without a clear discriminator.